tests: moved ad specific authentication test and created test_ad.py
